test: Send keys only to the active window
Instead of sending keys directly to a specific window, activate the
window, then send the keys to the active window.

This is work-around for an xdotool bug[0], which broke sending synthetic
keys, such as Ctrl+F, to a specific window.

The downside of this approach is that the user could break the test if
they switched to another window while the test is running. This is not a
big deal, because one should run automated tests in an isolated
environment (for example, under Xvfb) anyway.

This commit should be reverted once the xdotool bug is fixed.

[0] https://github.com/jordansissel/xdotool/issues/52